# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sidequestd API",
    "description": "Shared API for the Sidequestd web and mobile clients.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/feed": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "Reviews and activity from the accounts you follow",
        "description": "SPEC \u00a76.4, newest first.\n\nItems are a discriminated union \u2014 **branch on `type`**. A `review` carries\nthe whole review, rating and media and live like and comment counts included,\nso a row is interactive on arrival and tapping through to the detail page\ntells the client nothing new. A `backlog_activity` is the lighter-weight line\nSPEC \u00a76.11 blends inline: who moved which game to which list. A\n`recommended_review` is SPEC \u00a76.4's blend \u2014 a review the caller does *not*\nfollow the author of, carrying a `reason` for why it is here. Label it as\nsuch: it is in the union as its own type precisely so it cannot be rendered\nas though somebody they follow wrote it.\n\nThere is no separate `GET /activity`, on purpose. SPEC \u00a76.4 is explicit that\nHome is unified and that there is no Activity tab, so a second endpoint would\nbe a second thing to keep ordered against this one.\n\nAn empty page is the normal state for a new account rather than an error; the\nclient fills it from `/feed/suggestions`. Nothing is blended into an empty\nfeed \u2014 a caller who follows nobody gets the empty state SPEC \u00a76.4 describes,\nnot a page of recommendations dressed up as one.",
        "operationId": "read_feed_api_v1_feed_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From a previous page",
              "title": "Cursor"
            },
            "description": "From a previous page"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Results per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Results per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_Annotated_Union_FeedReviewItem__FeedActivityItem__FeedRecommendedItem___FieldInfo_annotation_NoneType__required_True__discriminator__type____"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feed/suggestions": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "What to show when the feed is empty",
        "description": "The empty state from SPEC \u00a76.4: somewhere to go when nobody is followed.\n\nA separate call rather than a field on the feed page, so the request that\nevery scroll makes does not also carry a section that is only ever read once.",
        "operationId": "read_suggestions_api_v1_feed_suggestions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedSuggestions"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BacklogStatus": {
        "type": "string",
        "enum": [
          "TO_BE_PLAYED",
          "PLAYING",
          "COMPLETED",
          "DROPPED"
        ],
        "title": "BacklogStatus"
      },
...
      "CursorPage_Annotated_Union_FeedReviewItem__FeedActivityItem__FeedRecommendedItem___FieldInfo_annotation_NoneType__required_True__discriminator__type____": {
        "properties": {
          "items": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/FeedReviewItem"
                },
                {
                  "$ref": "#/components/schemas/FeedActivityItem"
                },
                {
                  "$ref": "#/components/schemas/FeedRecommendedItem"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "backlog_activity": "#/components/schemas/FeedActivityItem",
                  "recommended_review": "#/components/schemas/FeedRecommendedItem",
                  "review": "#/components/schemas/FeedReviewItem"
                }
              }
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as ?cursor= for the next page; null when exhausted."
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CursorPage[Annotated[Union[FeedReviewItem, FeedActivityItem, FeedRecommendedItem], FieldInfo(annotation=NoneType, required=True, discriminator='type')]]"
      },
...
      "FeedActivityItem": {
        "properties": {
          "type": {
            "type": "string",
            "const": "backlog_activity",
            "title": "Type"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "The backlog item's id. Unique per `type`."
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At",
            "description": "When the status changed, newest first"
          },
          "actor": {
            "$ref": "#/components/schemas/UserPublic",
            "description": "Whose list moved"
          },
          "game": {
            "$ref": "#/components/schemas/GameSummary"
          },
          "status": {
            "$ref": "#/components/schemas/BacklogStatus",
            "description": "The list the game is on now"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "occurred_at",
          "actor",
          "game",
          "status"
        ],
        "title": "FeedActivityItem",
        "description": "Someone you follow moved a game between backlog lists (SPEC \u00a76.11).\n\nThe lightweight half of the unified Home from SPEC \u00a76.4 \u2014 \"Alex added *Elden\nRing* to Playing\", \"Sam completed *Hades*\" \u2014 which exists so Home still has a\npulse on a day when nobody you follow wrote a review. Reviews stay the hero\ncontent; this is a line, not a card.\n\nOnly the *current* status is here, not what it moved from. The event is\nderived from the backlog row itself rather than from an events table, and\nthat row remembers where a game is, not the route it took. \"Completed\n*Hades*\" is the whole of what SPEC \u00a76.11 asks for, and inventing a `from`\nfield the data cannot back would be worse than not having one."
      },
      "FeedRecommendedItem": {
        "properties": {
          "type": {
            "type": "string",
            "const": "recommended_review",
            "title": "Type"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "The review's id. Unique per `type`."
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At",
            "description": "When the review was written, newest first"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewSummary"
          },
          "reason": {
            "$ref": "#/components/schemas/RecommendationReason",
            "description": "What put this in front of the reader"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "occurred_at",
          "review",
          "reason"
        ],
        "title": "FeedRecommendedItem",
        "description": "A review from outside the follow graph, blended in (SPEC \u00a76.4).\n\nIts own member of the union rather than a `recommended: bool` on\n`FeedReviewItem`, because SPEC \u00a76.4 requires the blend to be \"clearly\ndistinguishable from pure follow feed\" and a flag is exactly the thing a\nclient forgets to read. A type it has to branch on cannot be missed, and the\nbranch is where the label and the \"why\" go.\n\nThe payload is the identical `ReviewSummary`, so a recommended row is as\ninteractive as any other \u2014 likeable, commentable, and by an author the reader\ncan follow on the spot. Nothing about the review is different; only how it\ngot here."
      },
      "FeedReviewItem": {
        "properties": {
          "type": {
            "type": "string",
            "const": "review",
            "title": "Type"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "The item's id \u2014 the review's, for this kind. Unique per `type`."
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At",
            "description": "What the feed is ordered by, newest first"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewSummary"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "occurred_at",
          "review"
        ],
        "title": "FeedReviewItem",
        "description": "A review from someone the viewer follows \u2014 the hero content of SPEC \u00a76.4.\n\nThe review is the full `ReviewSummary` every other surface returns, counters\nincluded, so a feed row renders a live like button and comment count without\na second request per item."
      },
      "FeedSuggestions": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/UserPublic"
            },
            "type": "array",
            "title": "Accounts",
            "description": "Public accounts worth following, most-followed first, excluding anyone the viewer already follows or has asked to."
          },
          "trending": {
            "items": {
              "$ref": "#/components/schemas/TrendingGame"
            },
            "type": "array",
            "title": "Trending",
            "description": "The materialised ranking from SPEC \u00a76.11. Empty until the trending job has run against some activity."
          }
        },
        "type": "object",
        "required": [
          "accounts",
          "trending"
        ],
        "title": "FeedSuggestions",
        "description": "What Home shows instead of a feed when there is nothing to show.\n\nSPEC \u00a76.4 asks for suggested accounts and trending content for a viewer who\nfollows nobody. Both sections are deliberately small: this is a prompt to go\nand follow someone, not a second Discover tab."
      },
...
      "GameSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/PlatformRef"
            },
            "type": "array",
            "title": "Platforms"
          },
          "release_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Year",
            "description": "Precomputed so every client renders the same year for a given date.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "cover_url",
          "release_date",
          "platforms",
          "release_year"
        ],
        "title": "GameSummary",
        "description": "The card shape used by browse and search results.\n\nDeliberately narrow: SPEC \u00a76.6 asks for cover art, release year and platform\non a result, and shipping the long `summary` text in a 20-item page would be\nmost of the payload for something no card renders."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlatformRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "PlatformRef"
      },
...
      "RecommendationReason": {
        "type": "string",
        "enum": [
          "recommended_game",
          "suggested_account"
        ],
        "title": "RecommendationReason",
        "description": "Why a recommended item is in front of you (SPEC \u00a76.4, \u00a76.5).\n\nSPEC \u00a76.4 asks for the blend to be \"clearly distinguishable from pure follow\nfeed\", and a separate `type` is what makes it distinguishable. This is the\nnext question a reader asks once they have noticed \u2014 *why this?* \u2014 and it is\na closed set rather than a sentence because the client writes the sentence\nand can translate it."
      },
...
      "ReviewSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "game": {
            "$ref": "#/components/schemas/GameSummary"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          },
          "playtime_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playtime Minutes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "comment_count": {
            "type": "integer",
            "title": "Comment Count"
          },
          "viewer_has_liked": {
            "type": "boolean",
            "title": "Viewer Has Liked",
            "description": "Whether the caller has liked this. Always false when signed out."
          },
          "media_count": {
            "type": "integer",
            "title": "Media Count"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url",
            "description": "What the profile grid tiles with: the first media thumbnail, falling back to the game's cover art (SPEC \u00a76.2)."
          },
          "stars": {
            "type": "number",
            "title": "Stars",
            "description": "The 0.5-5.0 value from SPEC \u00a76.3. Derived so clients cannot drift.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "game",
          "rating",
          "review_text",
          "playtime_minutes",
          "created_at",
          "updated_at",
          "like_count",
          "comment_count",
          "viewer_has_liked",
          "media_count",
          "thumbnail_url",
          "stars"
        ],
        "title": "ReviewSummary",
        "description": "A review as a feed row or a grid tile shows it.\n\nNo `from_attributes`: the counters are not columns, so this is assembled from\na `ReviewWithStats` rather than validated straight off the model."
      },
...
      "TrendingGame": {
        "properties": {
          "game": {
            "$ref": "#/components/schemas/GameSummary"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "window": {
            "$ref": "#/components/schemas/TrendingWindow",
            "description": "Rolling window the score was computed over"
          }
        },
        "type": "object",
        "required": [
          "game",
          "score",
          "window"
        ],
        "title": "TrendingGame",
        "description": "A trending entry: the card plus the score that ranked it (SPEC \u00a76.11)."
      },
      "TrendingWindow": {
        "type": "string",
        "enum": [
          "24h",
          "7d",
          "30d"
        ],
        "title": "TrendingWindow",
        "description": "Rolling windows the trending worker computes (SPEC \u00a76.11).\n\nClosed set rather than free text so a typo can't quietly return an empty\nranking; the durations live in `app.services.trending.TRENDING_WINDOWS`."
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "bio",
          "avatar_url",
          "is_private",
          "created_at"
        ],
        "title": "UserPublic",
        "description": "The profile shell visible to anyone, including non-followers of a private\naccount (SPEC \u00a76.7)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /auth/login",
        "scheme": "bearer"
      }
    }
  }
}